+++ Source/Podio .NET/Models/PodioResponse.cs
@@ -5,7 +5,7 @@ namespace PodioAPI.Models
     public class PodioResponse
     {
         public dynamic Body { get; set; }
-        public int Status { get; set; }
+        public long Status { get; set; }
         public Dictionary<string, string> Headers { get; set; }
         public string RequestUri { get; set; }
     }
